% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venn.R
\name{combiSet}
\alias{combiSet}
\title{Generates combinations separated by '&'

Generates all combinations or the elements in a vector.}
\usage{
combiSet(vec, low_order = TRUE)
}
\arguments{
\item{vec}{A vector with names}

\item{low_order}{Whether to return the low order combinations or only the full}
}
\description{
Generates combinations separated by '&'

Generates all combinations or the elements in a vector.
}
