% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venn.R
\name{customVenn}
\alias{customVenn}
\title{Euler diagram}
\usage{
customVenn(vennlist, main = "")
}
\arguments{
\item{vennlist}{A named list where each entry is a vector of the elements of the set}
}
\description{
Make an euler diagram, always works exactly up to 3 sets, then has to approximate so use equisurface venn diagram
}
