% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesets.R
\name{genesetsOverview}
\alias{genesetsOverview}
\alias{genesetsFromCluster}
\alias{compareClusters}
\title{Wrapper for CLICK cluster enrichment.}
\usage{
genesetsOverview(clusters, ...)

genesetsFromCluster(clusters, selection, ...)

compareClusters(zscores, clusters, cids)
}
\arguments{
\item{clusters}{CLICK cluster output with colums Cluster and Ensembl}

\item{selection}{Clusters on which gene set enrichment should be performed}

\item{zscores}{Z-normalised expression data with genes in rownames and samples in colnames.}

\item{cids}{Ids of the clusters to plot}
}
\description{
Wrapper for CLICK cluster enrichment.

Plot heatmap for multiple clusters
}
