% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kegg.R
\name{plotEnrichment}
\alias{plotEnrichment}
\title{Plot enrichment result}
\usage{
plotEnrichment(
  enriched_result,
  max_terms = 10,
  ylab = "KEGG pathway",
  title = paste("Enrichment score of", ylab),
  ref_enrichment = "",
  ref_rep = "one"
)
}
\arguments{
\item{enriched_result}{enrichResult}

\item{max_terms}{The maximum number of terms to display on the plot}

\item{ref_enrichment}{enrichResult for the universe set}

\item{ref_rep}{Representation of the reference enrichment, either the richFactor is computed using the reference set count as the denominator ('one'), or the richFactor of the reference set is displayed on the plot ('plot')}
}
\value{
The input with an extra column richFactor
}
\description{
Compute and plots the rich factor for the enriched pathways or terms.
}
