% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{plotPCA.prcomp}
\alias{plotPCA.prcomp}
\title{Set of functions to plot PCA results in tidyverse, 
Work with the prcomp objects from the stats package
Should use the generic function defined by BiocGenerics
Only prepares a ggplot object}
\usage{
plotPCA.prcomp(pca, components = 1:2)
}
\description{
Set of functions to plot PCA results in tidyverse, 
Work with the prcomp objects from the stats package
Should use the generic function defined by BiocGenerics
Only prepares a ggplot object
}
