\name{rBioHelp-package}
\alias{rBioHelp-package}
\alias{rBioHelp}
\docType{package}
\title{
\packageTitle{rBioHelp}
}
\description{
\packageDescription{rBioHelp}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{rBioHelp}
\packageIndices{rBioHelp}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{rBioHelp}

Maintainer: \packageMaintainer{rBioHelp}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
