% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpplot.R
\name{replace_bg}
\alias{replace_bg}
\title{Helper function to change the background for a gene}
\usage{
replace_bg(svg, gene, positions, colors)
}
\description{
Helper function to change the background for a gene
}
