% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmaps.R
\name{starHeatmap2}
\alias{starHeatmap2}
\title{Heatmap with stars for extra data}
\usage{
starHeatmap2(
  hm_values,
  data,
  encoding = "size",
  return_list = FALSE,
  range = -5:-1,
  hm_type = "Value",
  star_type = "log p-value",
  threshold = 0,
  ...
)
}
\arguments{
\item{hm_values}{Values for the colors of the heatmap tiles}

\item{data}{Values for the text (or number of stars) of the heatmap tiles. Should be all >=0 or all <=0.}

\item{encoding}{One of c('size', 'count', 'multi_count') for whether the value in data should be encoded by the size of one star or a number of stars on one or multiple lines}

\item{return_list}{If FALSE the heatmap is plotted with a legend, if TRUE returns a list with fields 'object' for the Heatmap object and 'annotation_legend_list' for the annotation corresponding to the encoding.}

\item{range}{The range of values that should be covered by the stars.}

\item{hm_type}{Name of the data type represented in the heatmap, used to name the legend.}

\item{star_type}{Name of the data type represented by the stars, used to name the legend.}

\item{threshold}{Threshold under which the stars are not plotted.}
}
\description{
Plot heatmap with stars in the tile to encode for another data layer (significance, effect size, etc)
}
