% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpplot.R
\name{WPplot}
\alias{WPplot}
\alias{WPmultiplot}
\title{Wrapper for wikiprofiler wpplot and wp_bgfill}
\usage{
WPplot(
  ftable,
  pathwayid = "WP4172",
  high = "red",
  low = "blue",
  legend = TRUE,
  legend_x = 0.001,
  legend_y = 0.94,
  strip_out = 0.01,
  fixed_range = F,
  lim = 5
)

WPmultiplot(
  ftable,
  pathwayid = "WP4172",
  multi = "Cell_line",
  high = "red",
  low = "blue",
  legend = TRUE,
  legend_x = 0.001,
  legend_y = 0.94,
  strip_out = 0.01,
  fixed_range = F,
  lim = 5
)
}
\arguments{
\item{ftable}{Either a tibble in long format with columns 'multi', Gene and beta containing the conditions, the gene symbols and the value to plot respectively; or a matrix with the conditions as column names and the genes as row names.}

\item{multi}{The name of the colum with the conditions}
}
\description{
Wrapper for wikiprofiler wpplot and wp_bgfill
}
