% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesets.R
\name{clusterSetsGraph}
\alias{clusterSetsGraph}
\title{Compute genesets enrichment for click clusters and plot the graph of gene sets relationships}
\usage{
clusterSetsGraph(
  clusters,
  clusters_enrichment,
  top_sets = 5,
  max_sets = 50,
  lwd_scaling = 10,
  node_label = "Description",
  node_size = "fixed"
)
}
\arguments{
\item{clusters}{Gene cluster assignement data.frame with columns Ensembl}

\item{clusters_enrichment}{A list with the enriched gene sets for each cluster}

\item{top_sets}{Number of top genesets to use for each CLICK cluster}

\item{max_sets}{Maximum number of genesets to plot}

\item{lwd_scaling}{Max line width, effectively the factor used to scale the line width in relation to the proportion of overlap in the gene sets. Use a high value if low overlaps are expected}

\item{cluster_downscaling}{CLICK clusters are way larger than genesets}
}
\description{
Compute genesets enrichment for click clusters and plot the graph of gene sets relationships
}
