% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesets.R
\name{nodeElement}
\alias{nodeElement}
\title{Build a named vector with all node names, so that it is valid for Rgraphviz::makeNodeAttrs}
\usage{
nodeElement(partial, graph, default_value)
}
\arguments{
\item{partial}{A named vector with node names as names and a value to set as value}

\item{graph}{The graph for which the attribute will be set}

\item{default_value, }{the value to set for the other nodes, uses the node name if default_value=='names' (useful for setting label).}
}
\description{
Build a named vector with all node names, so that it is valid for Rgraphviz::makeNodeAttrs
}
\examples{
\dontrun{ makeNodeAttrs(graph, fill=nodeElement(setNames(1='blue'), graph, 'red'), label=nodeElement(setNames(1='text'), graph, "names")) }
}
